requirements.txt
aniso8601==1.2.0
APScheduler==3.3.1
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.2.0
Automat==0.7.0
autopep8==1.4
bsddb3==6.2.6
cffi==1.11.5
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.2
django-allauth==0.37.1
django-debug-toolbar==1.10.1
django-extensions==2.1.2
django-filter==2.0.0
django-grappelli==2.11.1
django-import-export==1.0.1
django-otp==0.5.0
django-rest-auth==0.9.3
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.3.0
drfdocs==0.0.11
et-xmlfile==1.0.1
Flask==1.0.2
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.4
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.13
numpy==1.15.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.7
pandas==0.23.4
parsel==1.5.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
PyMySQL==0.7.11
pyOpenSSL==18.0.0
pypiwin32==223
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
pywin32==223
PyYAML==3.13
queuelib==1.5.0
redis==2.10.6
requests==2.13.0
requests-oauthlib==1.0.0
Scrapy==1.5.1
scrapy-deltafetch==1.2.1
scrapy-proxies==0.4
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==17.0.0
six==1.10.0
South==1.0.2
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
sqlparse==0.2.4
tablib==0.12.1
Twisted==18.7.0
typed-ast==1.1.0
tzlocal==1.3
unicodecsv==0.14.1
w3lib==1.19.0
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
